research(concept-dynamics): FINDINGS — rhythm is complex-recurrence; transformer commits, Mamba-1 cannot oscillate by construction
Sharpened the rhythm question from empirical to structural. FACT: oscillation capacity =
complex-valued temporal recurrence. Transformer = no recurrence -> no rhythm mechanism.
Mamba-1.4B state matrix A=-exp(A_log): all 3.1M eigenvalues real-negative (verified on
weights, range [-43657,-0.004], zero imag) -> pure exponential decay, NO oscillation
capacity by construction. EMPIRICAL: transformer concept signal 0/20 oscillation (commits;
slope -0.15, AR1 0.30, random null 0/20) -- validated instrument. HYPOTHESIS (marked):
rhythm is substrate-specific MECHANISM not substrate-independent FUNCTION -- brains route/
bind/multiplex WITH rhythm (coherence, theta-gamma 7+-2) because recurrent+analog;
transformers do the same functions via attention/positional-encoding/full-context, no
rhythm. => rhythm fundamental to RECURRENT cognition, not cognition. Falsifiable decider:
complex SSM (Mamba-3) should oscillate; is it FUNCTIONAL (ablate->routing breaks) or
epiphenomenal? Needs complex-SSM weights = next bet. (Also: ssm contrast dtype fix fp16->
fp32 for Mamba no-kernel path.)

+++ papers/concept-dynamics/measure_ssm_contrast.py
@@ -67,7 +67,7 @@ def main():
     for arch, mid in ARCHS:
         try:
             tok = AutoTokenizer.from_pretrained(mid)
-            mdl = AutoModelForCausalLM.from_pretrained(mid, dtype=torch.float16).to("cuda").eval()
+            mdl = AutoModelForCausalLM.from_pretrained(mid, dtype=torch.float32).to("cuda").eval()
             eos = tok.eos_token_id; eos = eos[0] if isinstance(eos, list) else eos
             pad = tok.pad_token_id if tok.pad_token_id is not None else eos
             nL = mdl.config.num_hidden_layers
@@ -78,8 +78,8 @@ def main():
             Hb = np.stack([last_resid(mdl, format_ids(tok, p), layer) for p in benign])
             v = Hh.mean(0) - Hb.mean(0)
             v = v / (np.linalg.norm(v) + 1e-9)
-            vt = torch.tensor(v, dtype=torch.float16, device=mdl.device)
-            R = torch.tensor(rng.standard_normal((N_RAND, len(v))), dtype=torch.float16, device=mdl.device)
+            vt = torch.tensor(v, dtype=torch.float32, device=mdl.device)
+            R = torch.tensor(rng.standard_normal((N_RAND, len(v))), dtype=torch.float32, device=mdl.device)
             R = R / R.norm(dim=1, keepdim=True)
 
             traj = []
